Add files via upload
the enco.text file is what needs to be encoded. I have uploaded the image files as well. The array is static as of now, hence the image files. It can be modified by adding a code segment that will populate the array with the images present inside the folder. The encoded binary version of the images will be created too along with the new image files . On running the decoder file, the decoded text will be generated in a file..